+++ lib/src/callbacks/threat_callback.dart
@@ -27,6 +27,7 @@ class ThreatCallback {
     this.onTimeSpoofing,
     this.onLocationSpoofing,
     this.onMultiInstance,
+    this.onDeviceBinding,
   });
 
   final VoidCallback? onRoot;
@@ -44,4 +45,5 @@ class ThreatCallback {
   final VoidCallback? onTimeSpoofing;
   final VoidCallback? onLocationSpoofing;
   final VoidCallback? onMultiInstance;
+  final VoidCallback? onDeviceBinding;
 }